Disable colorization only for R CMD check
Colorized test output was disabled for non-interactive tests because it
inserted non-printable characters into the output of `R CMD check`.  This patch
attempts to detect when R CMD check is in operation so that other
non-interactive methods of running tests, such as Rscript, can show how many
bugs you introduced since your last cup of coffee in fully colorized glory.

The fix relies on an apparently undocumented environment variable, `R_TESTS`
that `R CMD check` uses to supply .Rprofile information.  If the length of the
string stored in this variable is non-zero, colorized output is disabled.