compare.numeric() messages about difference in names and attributes
Previously, if all.equal() was not TRUE due to a difference like x
being named and y not being named, you would get a message from
compare() about the differences in the numeric values and nothing
about the names. That was very confusing to me since the differences
were all within the tolerance of given to all.equal(). Maybe it would
make sense to have compare.numeric() only detect difference above the
tolerance and call NextMethod() to detect attribute differences but I
thought I would try the simpler fix first, and also not change the
current behavior more than is necessary to avoid confusion.