Add view.on.update parameter for expect_equal_to_reference
It would be helpful if `expect_equal_to_reference` could allow test runners to view the new test results when they are being [re]created. It defaults to off, but it allows test authors to choose to force test runners to view the data. This probably needs some kind of extra checking so it doesn't run when testing is being done in a headless/batch mode, but I'm not sure what the right check to use would be.