Fix expect_output and capture_output encoding
We root of the problem is thar `capture_output` cannot
capture UTF-8 output, it is just not possible currently
with R. The output will always be in the native encoding.

If UTF-8 characters are outputted, then R automatically
converts them to the native encoding, using the equivalent
of `enc2native`.

`testthat` currently requires that the test files are
in UTF-8, and also that in `expect_output` UTF-8
characters are outputted. R will convert the UTF-8
characters to native encoding, and `testthat` does
the same now with the expected output, so they can
be compared on all platforms.

This commit:
* Adds an `encoding` argument to `read_lines`, to
  be able to mark the re-read output as `unknown`,
  which mean the native encoding. (It was erroneously
  marked as UTF-8 before.)
* Uses `encoding = "unknown"` in `expect_output`.
* Adds a test case for UTF-8 characters in `expect_output`.