Fix reporter test cases on non-UTF-8 platforms
There are two issues here. The first one is that we need to
make sure that we compare the strings in the same (native)
encoding. So we
* mark the actual output as in the native encoding,
  instead of (the default) UTF-8, in `read_lines`, and
* convert the reference output to the native encoding,
  from UTF-8.

The other issue is that we need to force Unicode (UTF-8)
characters in `cli`, otherwise we can't compare the output
with the reference output. This is because `cli` is smart
and avoids omitting UTF-8 characters on non-UTF-8 platforms
But the UTF-8 -> native conversion performed by R
is not the same as the fallback mechanism of `cli`, so
the output is different. By making sure that the `cli`
output is UTF-8, both strings will have the same
UTF-8 -> native conversion and they are comparable.