Fix encoding issues in output recording and comparison (#679)
* Fix expect_output and capture_output encoding

We root of the problem is thar `capture_output` cannot
capture UTF-8 output, it is just not possible currently
with R. The output will always be in the native encoding.

If UTF-8 characters are outputted, then R automatically
converts them to the native encoding, using the equivalent
of `enc2native`.

`testthat` currently requires that the test files are
in UTF-8, and also that in `expect_output` UTF-8
characters are outputted. R will convert the UTF-8
characters to native encoding, and `testthat` does
the same now with the expected output, so they can
be compared on all platforms.

This commit:
* Adds an `encoding` argument to `read_lines`, to
  be able to mark the re-read output as `unknown`,
  which mean the native encoding. (It was erroneously
  marked as UTF-8 before.)
* Uses `encoding = "unknown"` in `expect_output`.
* Adds a test case for UTF-8 characters in `expect_output`.

* Fix encoding in `expect_known_output`

See the previous commit about `expect_output`
for details. `expect_known_output` still
relies on the reference output having an UTF-8
encoding, but also makes sure that it is
converted to the native encoding for comparison.

* Fix reporter test cases on non-UTF-8 platforms

There are two issues here. The first one is that we need to
make sure that we compare the strings in the same (native)
encoding. So we
* mark the actual output as in the native encoding,
  instead of (the default) UTF-8, in `read_lines`, and
* convert the reference output to the native encoding,
  from UTF-8.

The other issue is that we need to force Unicode (UTF-8)
characters in `cli`, otherwise we can't compare the output
with the reference output. This is because `cli` is smart
and avoids omitting UTF-8 characters on non-UTF-8 platforms
But the UTF-8 -> native conversion performed by R
is not the same as the fallback mechanism of `cli`, so
the output is different. By making sure that the `cli`
output is UTF-8, both strings will have the same
UTF-8 -> native conversion and they are comparable.

* expect_known_output: warn for non-UTF-8 reference

This can easily happen if the output is recorded on
a non-UTF-8 platform. We also warn if the updated
output is not UTF-8.

This (with the previous commits) closes #672.