Expose chdir arguments in test functions
Sometimes you don't want to run tests in the environment
where those files are. For example users may not have write
access to an R package installation directory but still
want to run tests that require file IO. This commit
exposes the `chdir` option of `source_file` to test
runner functions. Can be disabled to run tests in the
current working directory.