Fix parallel tests (#1130)
* Remove unused arg from queue_setup()

* Fix parallel tests

The "none" loading strategy is not appropriate for
the subprocess, which does need to load the package
somehow.

* Always load helpers in the subprocess

* Parallel startup: call funcs from testthat

The startup code does not run inside the package
namespace, so we need to explicitly run it
within asNamespace("testthat").

* Set package name in subprocess

This is only needed in the parallel case, because
otherwise the test environment is created externally.

* Fix parallel teardown

callr does not work well with function objects from
packages, because it resets the environment of the
function to .GlobalEnv. Best is to pass an anonoymous
function.