Improve failure summaries
All reporters use issue_header() (extracted out of issue_summary()) which presents data in order of important: failure type, failure location, and test name. Failure type is coloured where possible, and if not broken types are displayed in upper case.

Backtraces for warnings are no longer displayed in CheckReporter to make it easier to see actual failures. I may revisit this decision in the future.

Fixes #1152