simplify patch per HW comments
Note we use the 'test_dir' deprecation warning in
'test_files' as that line may be triggered from either
'test_file' or 'test_dir'.  In the docs 'test_file' is
described as 'a variant of test_dir()...' so it seems
ok for 'test_file' to trigger a deprecation warning
with 'test_dir' in it.