Improve snapshotting cleanup in parallel tests (#1800)
We cannot forward all work of the snapshot reporter to the main process, because `expect_snapshot_helper()` uses the return value of the `take_snapshot()` method. So this must be called in the subprocess.

This is also good for performance, because the snapshot comparison code still runs in the subprocess, and we don't copy potentially large objects between processes.

The disadvantage is that we need slightly more complicated reporters, both in the subprocess and in the main process, to make sure that both processes are doing the right type of snapshot cleanup.

Closes #1797.

Simplifies and improves #1788 and #1793.