Add a reporter that emits test-related Open Telemetry spans.
This commit introduces a variant of the Check reporter that records Open
Telemetry traces for package tests. At present it records a parent span
for each context/file and child spans for each test case.

Span attributes are drawn from the semantic conventions [0].

It also makes some effort to work out the package's GitHub URL so that
we can set VCS-related attributes.

Note that this uses the existing session API from the `otel` package to
allow parallel testing, but this API has known issues.

Basic unit tests are included.

[0]: https://opentelemetry.io/docs/specs/semconv/registry/attributes/test/

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>