✨ Don't lose last snapshot if last nl is missing ✨
Because positron (and other editors) have an edit of trimming additional trailing newlines when saving. It probably would have been better to not write this in the first place, but I think we don't want to change that now as it will cause diffs to all snapshots, and only protects against a problem that happens rarely.

Fixes #2092