Fix parallel startup error message
Now that we read out stdout/stderr explicitly, we might read it
out before the error is reported from a failed worker startup.
Or if a task is a startup and not a test file (`path == NA`), then
we put the stdout/stderr aside and use it in the error message.