Skip snapshot on CRAN at end of test (#2218)
This ensures that the snapshots are not deleted. This also made me realise I missed the analgous update in `expect_snapshot_file()` and it's also clear that it should announce the file so it doesn't get deleted.

With these changes in place I can now run all the tests and ensure they pass on CRAN. This fixes #2214