Add `required` argument to expect_true
Adds `required` argument to `expect_true()` that accepts `c("one", "any", "all")`. I've held off implementing for `expect_false` until we make sure the implementation is sound for easier review.

When `required = "any"` we throw an error to avoid edges with numeric vectors that contain `1` which may be coerced to `TRUE`.

Fixes #1836.